backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
redis==5.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0
aiohttp==3.9.3
numpy==1.26.4
tenacity==8.2.3
python-multipart==0.0.6
jinja2==3.1.3

# Google Gemini AI
google-generativeai>=0.8.0

# Local Embeddings
sentence-transformers==2.3.1
